style: FRONTEND: implement short version of ids

+++ frontend/src/components/MainContentRightBar.jsx
@@ -6,6 +6,8 @@ import { sessionMetadataExtractor } from "../helpers/dataExtractors"
 import Map from "./singles/Map"
 import SessionMetadataTable from "./singles/SessionMetadataTable"
 
+import { short } from "../helpers/dataFormatters"
+
 const MainContentRightBar = ({session}) => {
   const [activeTab, setActiveTab] = useState('Events')
 
@@ -17,7 +19,7 @@ const MainContentRightBar = ({session}) => {
     <div className="main-right-bar">
 
       <div className="main-right-bar-session-details">
-        <h2>Session {`#${session.id}`}</h2>
+        <h2>Session {`#${short(session.id)}`}</h2>
         <Map session={session}/>
         <SessionMetadataTable session={session} /> 
       </div>
